+++ DependencyInjection/Configuration.php
@@ -7,6 +7,7 @@ use Symfony\Component\Config\Definition\ConfigurationInterface;
 
 final class Configuration implements ConfigurationInterface
 {
+    public const EXCEPTION_STRATEGY_CUSTOM = 'custom';
     public const EXCEPTION_STRATEGY_TO_HTTP_NOT_FOUND = 'to_http_not_found';
 
     public function getConfigTreeBuilder()
@@ -26,8 +27,42 @@ final class Configuration implements ConfigurationInterface
                 ->arrayNode('exceptions_strategy')
                     ->addDefaultsIfNotSet()
                     ->children()
-                        ->scalarNode('out_of_range_page')->defaultValue(self::EXCEPTION_STRATEGY_TO_HTTP_NOT_FOUND)->end()
-                        ->scalarNode('not_valid_current_page')->defaultValue(self::EXCEPTION_STRATEGY_TO_HTTP_NOT_FOUND)->end()
+                        ->scalarNode('out_of_range_page')
+                            ->defaultValue(self::EXCEPTION_STRATEGY_TO_HTTP_NOT_FOUND)
+                            ->validate()
+                                ->ifTrue(static function ($v) { return !\in_array($v, [self::EXCEPTION_STRATEGY_CUSTOM, self::EXCEPTION_STRATEGY_TO_HTTP_NOT_FOUND]); })
+                                ->then(static function ($v) {
+                                    @trigger_error(
+                                        sprintf(
+                                            'Setting the "babdev_pagerfanta.exceptions_strategy.out_of_range_page" configuration option to "%s" is deprecated since BabDevPagerfantaBundle 2.2, set the option to one of the allowed values: [%s]',
+                                            $v,
+                                            implode(', ', [self::EXCEPTION_STRATEGY_CUSTOM, self::EXCEPTION_STRATEGY_TO_HTTP_NOT_FOUND])
+                                        ),
+                                        E_USER_DEPRECATED
+                                    );
+
+                                    return $v;
+                                })
+                            ->end()
+                        ->end()
+                        ->scalarNode('not_valid_current_page')
+                            ->defaultValue(self::EXCEPTION_STRATEGY_TO_HTTP_NOT_FOUND)
+                            ->validate()
+                                ->ifTrue(static function ($v) { return !\in_array($v, [self::EXCEPTION_STRATEGY_CUSTOM, self::EXCEPTION_STRATEGY_TO_HTTP_NOT_FOUND]); })
+                                ->then(static function ($v) {
+                                    @trigger_error(
+                                        sprintf(
+                                            'Setting the "babdev_pagerfanta.exceptions_strategy.not_valid_current_page" configuration option to "%s" is deprecated since BabDevPagerfantaBundle 2.2, set the option to one of the allowed values: [%s]',
+                                            $v,
+                                            implode(', ', [self::EXCEPTION_STRATEGY_CUSTOM, self::EXCEPTION_STRATEGY_TO_HTTP_NOT_FOUND])
+                                        ),
+                                        E_USER_DEPRECATED
+                                    );
+
+                                    return $v;
+                                })
+                            ->end()
+                        ->end()
                     ->end()
                 ->end()
             ->end();