+++ Tests/DependencyInjection/ConfigurationTest.php
@@ -33,8 +33,25 @@ final class ConfigurationTest extends TestCase
     {
         $extraConfig = [
             'exceptions_strategy' => [
-                'out_of_range_page' => 'custom_handler',
-                'not_valid_current_page' => 'custom_handler',
+                'out_of_range_page' => Configuration::EXCEPTION_STRATEGY_CUSTOM,
+                'not_valid_current_page' => Configuration::EXCEPTION_STRATEGY_CUSTOM,
+            ],
+        ];
+
+        $config = (new Processor())->processConfiguration(new Configuration(), [$extraConfig]);
+
+        $this->assertEquals(
+            array_merge(self::getBundleDefaultConfig(), $extraConfig),
+            $config
+        );
+    }
+
+    public function testConfigWithDeprecatedExceptionsStrategy(): void
+    {
+        $extraConfig = [
+            'exceptions_strategy' => [
+                'out_of_range_page' => 'deprecated',
+                'not_valid_current_page' => 'deprecated',
             ],
         ];